Loader: move from loading sectionheader to loading program headers
This change is required to optimize memory accesses, since two program headers marked PT_LOAD do not share any memory pages. This is not the case for section headers.
We lose a bit in debugability, since we no longer show the names of the sections when we crash. But this can be readded later if necessary anyway.