Kernel: let the kernel execute syscalls.
We cannot execute syscalls as regular instructions, because they
may block the current thread. Getting control back after the thread
is unblocked looks very complicated to do. Let's instead have the
kernel take back control of the thread by syncing the state and
executing the syscall directly on the saved state.
This way, we can also block easily and execute the end of the syscall
later within the blocker logic.

I really want to keep Cpu and VM as local variables only, and giving
the kernel anytime-access to them seems like an easy way out but also
a bad idea.