Mmu: perform direct page access rather than using regions
This is the optimization unlocked by the previous commit. It yields
a nice 2-4% speedup on most test instances, but it scales to a 15%
speedup on heavy memcpy workloads.

This optimization is somewhat unsafe in release mode, because we
are no longer checking out of bound reads/writes as we did in the
region. This is particularly annoying for cross-region 128-bit reads
or writes, which we have already seen...

This motivates even more a change to a non-memory owning region
architectures, where all memory is allocated in one contiguous
virtual memory range.