Mmu: transition to a flat memory model.
Rather than having Mmu::Regions own the memory, we have the Mmu reserve
a large chunk of memory on the Host close to 0. This area is usually empty
(as long as the emulator is compiled as a dynamic object file). Regions now
become views into that memory, and we mprotect sub-regions according to the
Mmu::Regions that are created. This prototype seems to work.